CORE-17955 - Merged the state creation with the state update inside the ClaimStateStore implementation.
