+++ components/ledger/ledger-utxo-token-cache/src/main/kotlin/net/corda/ledger/utxo/token/cache/factories/TokenCacheEventProcessorFactoryImpl.kt
@@ -54,7 +54,7 @@ class TokenCacheEventProcessorFactoryImpl(
             eventConverter,
             entityConverter,
             tokenPoolCacheManager,
-            ClaimStateStoreCacheImpl(stateManager, serialization, claimStateStoreFactory, clock),
+            ClaimStateStoreCacheImpl(claimStateStoreFactory),
             externalEventResponseFactory,
             tokenSelectionMetrics
         )
