+++ components/ledger/ledger-utxo-token-cache/src/main/kotlin/net/corda/ledger/utxo/token/cache/services/ClaimStateStoreFactory.kt
@@ -3,5 +3,5 @@ package net.corda.ledger.utxo.token.cache.services
 import net.corda.ledger.utxo.token.cache.entities.TokenPoolKey
 
 interface ClaimStateStoreFactory {
-    fun create(key: TokenPoolKey, storedPoolClaimState: StoredPoolClaimState): ClaimStateStore
+    fun create(key: TokenPoolKey): ClaimStateStore
 }
