+++ components/ledger/ledger-utxo-token-cache/src/test/kotlin/net/corda/ledger/utxo/token/cache/impl/services/ClaimStateStoreCacheImplTest.kt
@@ -2,14 +2,12 @@ package net.corda.ledger.utxo.token.cache.impl.services
 
 import net.corda.ledger.utxo.token.cache.impl.POOL_KEY
 import net.corda.ledger.utxo.token.cache.impl.TOKEN_POOL_CACHE_STATE
-import net.corda.ledger.utxo.token.cache.services.StoredPoolClaimState
 import net.corda.ledger.utxo.token.cache.services.ClaimStateStore
 import net.corda.ledger.utxo.token.cache.services.ClaimStateStoreCacheImpl
 import net.corda.ledger.utxo.token.cache.services.ClaimStateStoreFactory
 import net.corda.ledger.utxo.token.cache.services.TokenPoolCacheStateSerialization
 import net.corda.libs.statemanager.api.State
 import net.corda.libs.statemanager.api.StateManager
-import net.corda.utilities.time.Clock
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.any
@@ -25,31 +23,24 @@ class ClaimStateStoreCacheImplTest {
     private val stateManager = mock<StateManager>()
     private val serialization = mock<TokenPoolCacheStateSerialization>()
     private val claimStateStoreFactory = mock<ClaimStateStoreFactory>()
-    private val clock = mock<Clock>().apply { whenever(instant()).thenReturn(now) }
 
-    private val target = ClaimStateStoreCacheImpl(stateManager, serialization, claimStateStoreFactory, clock)
+    private val target = ClaimStateStoreCacheImpl(claimStateStoreFactory)
 
     @Test
     fun `get when no existing state in store then create default state`() {
         val newClaimStateStore = mock<ClaimStateStore>()
         val stateBytes = "data".toByteArray()
         whenever(stateManager.get(any())).thenReturn(mapOf())
-        whenever(claimStateStoreFactory.create(any(), any())).thenReturn(newClaimStateStore)
+        whenever(claimStateStoreFactory.create(any())).thenReturn(newClaimStateStore)
         whenever(serialization.serialize(any())).thenReturn(stateBytes)
 
         val expectedState = TOKEN_POOL_CACHE_STATE
         val expectedStoredState = State(POOL_KEY.toString(), stateBytes, modifiedTime = now)
 
-        val expectedStoredPoolClaimState = StoredPoolClaimState(
-            State.VERSION_INITIAL_VALUE,
-            POOL_KEY,
-            expectedState
-        )
-
         val result = target.get(POOL_KEY)
         assertThat(result).isEqualTo(newClaimStateStore)
 
-        verify(claimStateStoreFactory).create(eq(POOL_KEY), eq(expectedStoredPoolClaimState))
+        verify(claimStateStoreFactory).create(eq(POOL_KEY))
         verify(stateManager).create(eq(listOf(expectedStoredState)))
     }
 
@@ -62,19 +53,13 @@ class ClaimStateStoreCacheImplTest {
 
         whenever(stateManager.get(any())).thenReturn(mapOf(POOL_KEY.toString() to stateRecord))
         whenever(serialization.deserialize(any())).thenReturn(expectedState)
-        whenever(claimStateStoreFactory.create(any(), any())).thenReturn(newClaimStateStore)
-
-        val expectedStoredPoolClaimState = StoredPoolClaimState(
-            stateRecord.version,
-            POOL_KEY,
-            expectedState
-        )
+        whenever(claimStateStoreFactory.create(any())).thenReturn(newClaimStateStore)
 
         val result = target.get(POOL_KEY)
         assertThat(result).isEqualTo(newClaimStateStore)
 
         verify(serialization).deserialize(stateBytes)
-        verify(claimStateStoreFactory).create(eq(POOL_KEY), eq(expectedStoredPoolClaimState))
+        verify(claimStateStoreFactory).create(eq(POOL_KEY))
     }
 
     @Test
@@ -83,14 +68,14 @@ class ClaimStateStoreCacheImplTest {
         val stateBytes = "data".toByteArray()
         whenever(stateManager.get(any())).thenReturn(mapOf())
         whenever(serialization.serialize(any())).thenReturn(stateBytes)
-        whenever(claimStateStoreFactory.create(any(), any())).thenReturn(newClaimStateStore)
+        whenever(claimStateStoreFactory.create(any())).thenReturn(newClaimStateStore)
 
         val result1 = target.get(POOL_KEY)
         assertThat(result1).isEqualTo(newClaimStateStore)
 
         val result2 = target.get(POOL_KEY)
-        verify(claimStateStoreFactory, times(1)).create(any(), any())
+        verify(claimStateStoreFactory, times(1)).create(any())
         assertThat(result2).isSameAs(result1)
-        verify(claimStateStoreFactory, times(1)).create(any(), any())
+        verify(claimStateStoreFactory, times(1)).create(any())
     }
 }
